weekend-01/requirements.txt
pick
openai
python-dotenv